Version 0.0.0.5 Code
Includes the changes to create the configuration automatically, and now contains framework for the updater system. (Not fully implemented) There are other minor adjustments and will begin optimizing and adjusting other parts that were added for initial builds GPU Mining code is still being worked on, but no changes in this release.